# src/NJobDesk.AspNetCore/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "NJobDesk API",
    "version": "v1"
  },
  "paths": {
...
    "/njobdesk/api/v1/executions": {
      "get": {
        "tags": [
          "Executions"
        ],
        "operationId": "getExecutions",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "provider",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jobId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jobName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ExecutionStatus"
            }
          },
          {
            "name": "fromUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "toUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionModelPagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/njobdesk/api/v1/executions/{id}/logs": {
      "get": {
        "tags": [
          "Executions"
        ],
        "operationId": "getExecutionLogs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExecutionLogModel"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/njobdesk/api/v1/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "operationId": "getJobs",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "provider",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSummaryModelPagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/njobdesk/api/v1/jobs/{id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "operationId": "getJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetailModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "operationId": "deleteJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/njobdesk/api/v1/jobs/{id}/trigger": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "operationId": "triggerJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetailModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/njobdesk/api/v1/jobs/{id}/pause": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "operationId": "pauseJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetailModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/njobdesk/api/v1/jobs/{id}/resume": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "operationId": "resumeJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetailModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/njobdesk/api/v1/triggers/{id}/pause": {
      "post": {
        "tags": [
          "Triggers"
        ],
        "operationId": "pauseTrigger",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/njobdesk/api/v1/triggers/{id}/resume": {
      "post": {
        "tags": [
          "Triggers"
        ],
        "operationId": "resumeTrigger",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/njobdesk/api/v1/triggers/{id}/reset-error": {
      "post": {
        "tags": [
          "Triggers"
        ],
        "operationId": "resetTriggerError",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/njobdesk/api/v1/triggers/{id}": {
      "delete": {
        "tags": [
          "Triggers"
        ],
        "operationId": "unscheduleTrigger",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/njobdesk/api/v1/triggers/{id}/schedule": {
      "put": {
        "tags": [
          "Triggers"
        ],
        "operationId": "rescheduleTrigger",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ExecutionLogLevel": {
        "enum": [
          "Trace",
          "Debug",
          "Information",
          "Warning",
          "Error",
          "Critical"
        ],
        "type": "string"
      },
      "ExecutionLogModel": {
        "required": [
          "category",
          "level",
          "message",
          "timestampUtc"
        ],
        "type": "object",
        "properties": {
          "timestampUtc": {
            "type": "string",
            "format": "date-time"
          },
          "level": {
            "$ref": "#/components/schemas/ExecutionLogLevel"
          },
          "category": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "exception": {
            "type": "string",
            "nullable": true
          },
          "properties": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExecutionModel": {
        "required": [
          "id",
          "jobId",
          "jobName",
          "providerKey",
          "recovering",
          "schedulerInstanceId",
          "startedUtc",
          "state"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "providerKey": {
            "type": "string"
          },
          "jobId": {
            "type": "string"
          },
          "jobName": {
            "type": "string"
          },
          "jobGroup": {
            "type": "string",
            "nullable": true
          },
          "triggerName": {
            "type": "string",
            "nullable": true
          },
          "startedUtc": {
            "type": "string",
            "format": "date-time"
          },
          "finishedUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "durationMs": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "state": {
            "$ref": "#/components/schemas/ExecutionStatus"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "schedulerInstanceId": {
            "type": "string"
          },
          "recovering": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ExecutionModelPagedResult": {
        "required": [
          "items",
          "total"
        ],
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExecutionModel"
            }
          }
        },
        "additionalProperties": false
      },
      "ExecutionStatus": {
        "enum": [
          "Running",
          "Succeeded",
          "Failed",
          "Vetoed"
        ],
        "type": "string"
      },
      "JobDetailModel": {
        "required": [
          "job",
          "recentExecutions",
          "triggers"
        ],
        "type": "object",
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobSummaryModel"
          },
          "triggers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TriggerModel"
            }
          },
          "recentExecutions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExecutionModel"
            }
          }
        },
        "additionalProperties": false
      },
      "JobState": {
        "enum": [
          "None",
          "Normal",
          "Paused",
          "Blocked",
          "Error",
          "Complete"
        ],
        "type": "string"
      },
      "JobSummaryModel": {
        "required": [
          "capabilities",
          "id",
          "isSystemJob",
          "name",
          "providerKey",
          "state",
          "triggerCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "providerKey": {
            "type": "string"
          },
          "group": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "jobType": {
            "type": "string",
            "nullable": true
          },
          "durable": {
            "type": "boolean",
            "nullable": true
          },
          "concurrentExecutionDisallowed": {
            "type": "boolean",
            "nullable": true
          },
          "triggerCount": {
            "type": "integer",
            "format": "int32"
          },
          "scheduleSummary": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "nextFireTimeUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "previousFireTimeUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isSystemJob": {
            "type": "boolean"
          },
          "capabilities": {
            "$ref": "#/components/schemas/SchedulerCapabilities"
          }
        },
        "additionalProperties": false
      },
      "JobSummaryModelPagedResult": {
        "required": [
          "items",
          "total"
        ],
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobSummaryModel"
            }
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "RescheduleRequestModel": {
        "required": [
          "cronExpression"
        ],
        "type": "object",
        "properties": {
          "cronExpression": {
            "type": "string"
          },
          "timeZoneId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SchedulerCapabilities": {
        "required": [
          "delete",
          "groups",
          "history",
          "interrupt",
          "pause",
          "runLogs",
          "scheduleEditing",
          "triggerNow",
          "triggers"
        ],
        "type": "object",
        "properties": {
          "triggerNow": {
            "type": "boolean"
          },
          "pause": {
            "type": "boolean"
          },
          "scheduleEditing": {
            "type": "boolean"
          },
          "delete": {
            "type": "boolean"
          },
          "groups": {
            "type": "boolean"
          },
          "triggers": {
            "type": "boolean"
          },
          "history": {
            "type": "boolean"
          },
          "runLogs": {
            "type": "boolean"
          },
          "interrupt": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "TriggerModel": {
        "required": [
          "id",
          "name",
          "startTimeUtc",
          "state",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "group": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/TriggerType"
          },
          "cronExpression": {
            "type": "string",
            "nullable": true
          },
          "cronSummary": {
            "type": "string",
            "nullable": true
          },
          "timeZoneId": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "nextFireTimeUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "previousFireTimeUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "startTimeUtc": {
            "type": "string",
            "format": "date-time"
          },
          "endTimeUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "misfireInstruction": {
            "type": "string",
            "nullable": true
          },
          "priority": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TriggerType": {
        "enum": [
          "Cron",
          "Simple",
          "Other"
        ],
        "type": "string"
      }
    }
  }
}